fix: lex operator combinations separately (#22)
* fix: lex operator combinations separately

This commit fixes a bug that would cause the lexer to parse
possibly-ambigious operator combinations as invalid. For instance,
"-(" may be a Minus and LParen symbol, but the lexer would treat the
combination as an invalid symbol. To avoid this, the lexer now consumes
tokens incrementally and falls back on the first-valid operator
combination discovered.

Closes #16.

* fixup! fix: lex operator combinations separately