Add support for lexing chars (#49)
* Add support for lexing chars

This commits adds the `Char` and `CharBound` symbols, and support for
tokenizing chars. Chars that do not contain exactly one character are
considered invalid. Escaped characters are also handled.

This commit also backfills a fix for proper spans of escaped characters,
which can span more than one character (as in the case of "\t") but were
previously recorded as only spanning one.

Part of #38

* fixup! Add support for lexing chars